Rename schema package and files to dbinfo
