+++ go/cmd/root.go
@@ -37,7 +37,7 @@ func Execute() {
 	root.AddCommand(testerCmd())
 	root.AddCommand(tracerCmd())
 	root.AddCommand(keysCmd())
-	root.AddCommand(schemaCmd())
+	root.AddCommand(dbinfoCmd())
 	root.AddCommand(transactionsCmd())
 
 	err := root.Execute()
